+++ TabletopTweaks/NewContent/Feats/QuickChannel.cs
@@ -20,7 +20,7 @@ namespace TabletopTweaks.NewContent.Feats {
             var QuickChannel = Helpers.CreateBlueprint<BlueprintFeature>("QuickChannel", bp => {
                 bp.SetName("Quick Channel");
                 bp.SetDescription("You may channel energy as a move action by spending 2 daily uses of that ability.");
-                bp.m_Icon = ExtraChannel.Icon;
+                bp.m_Icon = SelectiveChannel.Icon;
                 bp.Ranks = 1;
                 bp.ReapplyOnLevelUp = true;
                 bp.IsClassFeature = true;