+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Version 1.6.0
 * dispel magic greater now correctly has a cap on how many buffs it can remove (1/4 CL)
-* added channeler of the Unknown
+* added new cleric archetype
+    * a fallen cleric archetype that has a unique domain selection, can convert all spells into domain spells, and gets a unique form of channeling that damages ALL things
+* added improved channel feat
+* added quick channel feat
 
 ## Version 1.5.2
 * preapplied buffs will now be added at the correct CL, this will result in many buffs being easier to dispell